move parseMethodName func to internal/grpcutil
