+++ internal/grpcutil/method.go
@@ -16,18 +16,17 @@
  *
  */
 
-package binarylog
+package grpcutil
 
 import (
 	"errors"
 	"strings"
 )
 
-// parseMethodName splits service and method from the input. It expects format
+// ParseMethod splits service and method from the input. It expects format
 // "/service/method".
 //
-// TODO: move to internal/grpcutil.
-func parseMethodName(methodName string) (service, method string, _ error) {
+func ParseMethod(methodName string) (service, method string, _ error) {
 	if !strings.HasPrefix(methodName, "/") {
 		return "", "", errors.New("invalid method name: should start with /")
 	}
